bug :: 버그 수정

+++ src/main/kotlin/xquare/app/xquareinfra/domain/container/adapter/dto/response/GetContainerDeployHistoryResponse.kt
@@ -10,7 +10,7 @@ data class DeployHistoryResponse(
     val scheduledDate: Long,
     val commitMessage: String,
     val stages: List<StageStatus>,
-    val pipelineCounter: Long,
+    val pipelineCounter: Int,
     val pipelineName: String
 )
